agent/fullstack/pyproject.toml
[project]
name = "neon-fullstack-agent"
version = "0.1.0"
description = "Opinionated Tanstack codegen agent."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic[bedrock]>=0.49.0",
    "anyio>=4.9.0",
    "dagger-io>=0.17.1",
    "fastapi>=0.115.12",
    "uvicorn==0.34.0",
]

[dependency-groups]
google = [
    "google-genai>=1.7.0",
]
openai = [
    "openai>=1.68.2",
]
